Have generator create application.js file if not found
Fixes #78
Generator used to put a setup error message when it did not find
application.js. Now, it will simply just create one with the
necessary data in it. Note that we don't generate a turbolinks
sprockets requirement if we create the file for the user since
this is not really related to implementing React on Rails.
Users should add this manually if they want turbolinks.