Fixes turbolinks loading issue
Code was checking that turbolinks was installed when it was not yet
because some setups load turbolinks after the bundles. The changes to
the code will check if turbolinks is installed after the page loaded
event fires.

Code was also added to allow easy debugging of Turbolinks, which should
be useful when v5 of Turbolinks is released shortly.

Details of how to configure Turbolinks with troubleshooting were added
to docs/additional_reading/turbolinks.md.

Also:

* Changed generation of the examples apps to /tmp/examples. This allows
  one to configure editors to ignore /tmp. Ignoring /examples is
  problematic as one deletes the /examples directory frequently.