Optimize efficiency of EnsureAssetsCompiled functionality
This commit changes the EnsureAssetsCompiled feature that ensures
RSpec tests are run with the latest webpack bundles. Previously,
webpack bundles would be rebuilt every test run, whether or not
they actually need to be. While one could around this by running
webpack in the background to rebuild the static assets, doing so
required users to name their webpack build script exactly correct
as the technique relied on a `pgrep`.

Instead, the functionality checks the mdate on each generated webpack
bundle/asset against the mdate of every file inside of the `client`
directory, which is all encapsulated by the standard library method
`FileUtils.uptodate?` and is very performant. This eliminates the
need to check for an existing webpack process because the webpack
bundles will be up to date anyway if the latter is running.

This required adding some fixtures for proper testing. These can be
found in `spec/react_on_rails/fixtures/webpack_assets`.