Justin808/updates for hmr with webpacker v4 (#1240)
* Added same_bundle_for_client_and_server

Added configuration option `same_bundle_for_client_and_server` with default `false` because

1. Production applications would typically have a server bundle that differs from the client bundle
2. This change only affects trying to use HMR with react_on_rails with rails/webpacker.

The previous behavior was to always go to the webpack-dev-server for the server bundle if the
webpack-dev-server was running _and_ the server bundle was found in the `manifest.json`.

If you are using the **same bundle for client and server rendering**, then set this configuration option
to `true`.

Also, remove the requirement that the server bundle have the default deployment
path. A subdirectory can be specified for the server_bundle_js_file.